# Prueba_FullStack_DilanRondon_Credyty/App.API/App.Servicios/DTOs/FacturaDTO.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace App.Servicios.DTOs
{
    public class FacturaDTO
    {
        public int FacturaID { get; set; }
        public int IngresoID { get; set; }
        public string? NumeroFactura { get; set; }
    }
}