# Prueba_FullStack_DilanRondon_Credyty/Swagger_API.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "API Parqueadero",
    "version": "v1"
  },
  "paths": {
    "/api/Autenticacion/adicionar": {
      "post": {
        "tags": [
          "Autenticacion"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsuarioDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UsuarioDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UsuarioDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Autenticacion/login": {
      "post": {
        "tags": [
          "Autenticacion"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AutenticacionDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AutenticacionDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AutenticacionDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Factura/facturas": {
      "get": {
        "tags": [
          "Factura"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Factura/factura/{id}": {
      "get": {
        "tags": [
          "Factura"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Factura/factura": {
      "post": {
        "tags": [
          "Factura"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FacturaDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FacturaDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FacturaDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Parqueadero/ingresos": {
      "get": {
        "tags": [
          "Parqueadero"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Parqueadero/ingreso/{id}": {
      "get": {
        "tags": [
          "Parqueadero"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Parqueadero"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParqueaderoDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ParqueaderoDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ParqueaderoDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Parqueadero"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Parqueadero/ingreso": {
      "post": {
        "tags": [
          "Parqueadero"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParqueaderoDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ParqueaderoDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ParqueaderoDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Tarifa/tarifas": {
      "get": {
        "tags": [
          "Tarifa"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TarifaDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TarifaDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TarifaDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Tarifa/tarifa/{id}": {
      "get": {
        "tags": [
          "Tarifa"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TarifaDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TarifaDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TarifaDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/Tarifa": {
      "post": {
        "tags": [
          "Tarifa"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TarifaDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TarifaDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TarifaDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Tarifa/{id}": {
      "put": {
        "tags": [
          "Tarifa"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TarifaDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TarifaDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TarifaDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Tarifa"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Vehiculo": {
      "get": {
        "tags": [
          "Vehiculo"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VehiculoDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VehiculoDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VehiculoDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Vehiculo"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehiculoDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/VehiculoDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/VehiculoDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Vehiculo/{id}": {
      "get": {
        "tags": [
          "Vehiculo"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/VehiculoDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehiculoDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/VehiculoDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Vehiculo"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehiculoDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/VehiculoDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/VehiculoDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Vehiculo"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AutenticacionDTO": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "clave": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "FacturaDTO": {
        "type": "object",
        "properties": {
          "facturaID": {
            "type": "integer",
            "format": "int32"
          },
          "ingresoID": {
            "type": "integer",
            "format": "int32"
          },
          "numeroFactura": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ParqueaderoDTO": {
        "type": "object",
        "properties": {
          "ingresoID": {
            "type": "integer",
            "format": "int32"
          },
          "vehiculoID": {
            "type": "integer",
            "format": "int32"
          },
          "horaIngreso": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "horaSalida": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "valorPagado": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "descuentoAplicado": {
            "type": "boolean",
            "nullable": true
          },
          "tipoVehiculo": {
            "type": "string",
            "nullable": true
          },
          "placa": {
            "type": "string",
            "nullable": true
          },
          "documentoIdentidad": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "TarifaDTO": {
        "type": "object",
        "properties": {
          "tarifaID": {
            "type": "integer",
            "format": "int32"
          },
          "tipoVehiculo": {
            "type": "string",
            "nullable": true
          },
          "precioPorMinuto": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "UsuarioDTO": {
        "type": "object",
        "properties": {
          "usuarioID": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "documentoIdentidad": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "clave": {
            "type": "string",
            "nullable": true
          },
          "fechaCreacion": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "VehiculoDTO": {
        "type": "object",
        "properties": {
          "vehiculoID": {
            "type": "integer",
            "format": "int32"
          },
          "tipoVehiculo": {
            "type": "string",
            "nullable": true
          },
          "placa": {
            "type": "string",
            "nullable": true
          },
          "documentoIdentidad": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Ingrese el token en el siguiente formato: Bearer {token}",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ]
}